Add separate component for indicator
