+++ src/platform-apps/channels/channel-view.test.tsx
@@ -117,25 +117,4 @@ describe('ChannelView', () => {
 
     expect(wrapper.find(Waypoint).exists()).toBe(false);
   });
-
-  it('renders indicator when new messages arrive', () => {
-    const wrapper = subject({ hasNewMessage: 0 });
-
-    expect(wrapper.find('.channel-view__newMessage').exists()).toBe(false);
-
-    wrapper.setProps({ hasNewMessage: 2 });
-
-    expect(wrapper.find('.channel-view__newMessage').exists()).toBe(true);
-  });
-
-  it('it should not renders indicator when click on it', () => {
-    const closeIndicatorSpy = jest.fn();
-    const wrapper = subject({ hasNewMessage: 3, closeIndicator: closeIndicatorSpy });
-
-    expect(wrapper.find('.channel-view__newMessage').exists()).toBe(true);
-
-    wrapper.find('.channel-view__newMessage-bar').simulate('click');
-
-    expect(wrapper.prop('scrollToBottom')).toHaveBeenCalled;
-  });
 });
