+++ src/platform-apps/channels/channel-view.tsx
@@ -5,6 +5,7 @@ import moment from 'moment';
 import { Message as MessageModel } from '../../store/messages';
 import { Message } from './message';
 import InvertedScroll from '../../components/inverted-scroll';
+import IndicatorMessage from '../../components/indicator_message';
 
 interface ChatMessageGroups {
   [date: string]: MessageModel[];
@@ -92,39 +93,11 @@ export class ChannelView extends React.Component<Properties> {
   render() {
     return (
       <div className='channel-view'>
-        {this.props.hasNewMessage > 0 && (
-          <div className='channel-view__newMessage'>
-            <button
-              type='button'
-              className='channel-view__newMessage-bar'
-              aria-label='Jump to last unread message'
-              onClick={this.scrollToBottom}
-            >
-              <span className='channel-view__newMessage-bar-text'>{this.props.hasNewMessage} new messages</span>
-            </button>
-            <button
-              type='button'
-              className='channel-view__newMessage-alt'
-              onClick={this.scrollToBottom}
-            >
-              Mark As Read
-              <svg
-                aria-hidden='true'
-                role='img'
-                width='24'
-                height='24'
-                viewBox='0 0 24 24'
-              >
-                <path
-                  fill='currentColor'
-                  fill-rule='evenodd'
-                  clip-rule='evenodd'
-                  d='M12.291 5.70697L15.998 9.41397L21.705 3.70697L20.291 2.29297L15.998 6.58597L13.705 4.29297L12.291 5.70697ZM1.99805 7H11.088C11.564 9.837 14.025 12 16.998 12V18C16.998 19.103 16.102 20 14.998 20H8.33205L2.99805 24V20H1.99805C0.894047 20 -0.00195312 19.103 -0.00195312 18V9C-0.00195312 7.897 0.894047 7 1.99805 7Z'
-                ></path>
-              </svg>
-            </button>
-          </div>
-        )}
+        <IndicatorMessage
+          hasNewMessage={this.props.hasNewMessage}
+          closeIndicator={this.props.closeIndicator}
+          scrollToBottom={this.scrollToBottom}
+        />
         <InvertedScroll className='channel-view__inverted-scroll'>
           <div className='channel-view__name'>
             <h1>Welcome to #{this.props.name}</h1>
