+++ src/platform-apps/channels/styles.scss
@@ -17,61 +17,6 @@
     padding: 100px 0 20px 20px;
     box-sizing: border-box;
 
-    &__newMessage {
-      display: flex;
-      justify-content: space-between;
-      position: absolute;
-      left: 16px;
-      right: 16px;
-      top: 0;
-      align-items: center;
-      z-index: 2;
-      min-height: 24px;
-      text-transform: capitalize;
-      -webkit-transition: opacity 0.15s;
-      transition: opacity 0.15s;
-      opacity: 0.95;
-      cursor: pointer;
-      border-radius: 0 0 8px 8px;
-      background-color: hsl(235, calc(1 * 85.6%), 64.7%);
-      box-shadow: 0 1px 0 hsla(240, calc(var(--saturation-factor, 1) * 11.1%), 1.8%, 0.2),
-        0 1.5px 0 hsla(240, calc(var(--saturation-factor, 1) * 7.7%), 2.5%, 0.05),
-        0 2px 0 hsla(240, calc(var(--saturation-factor, 1) * 11.1%), 1.8%, 0.05);
-      &-bar {
-        cursor: pointer;
-        background: none;
-        border: 0;
-        text-align: left;
-        flex: 1 1 auto;
-        display: block;
-        position: relative;
-        white-space: nowrap;
-        text-overflow: ellipsis;
-        overflow: hidden;
-        &-text {
-          outline: 0;
-        }
-      }
-
-      &-alt {
-        cursor: pointer;
-        background: none;
-        border: 0;
-        flex: 0 0 auto;
-        font-weight: 600;
-        position: relative;
-        display: contents;
-        svg {
-          display: block;
-          position: relative;
-          top: -1px;
-          width: 16px;
-          height: 16px;
-          margin-left: 8px;
-        }
-      }
-    }
-
     &__name {
       padding: 10px 0;
       h1 {
